bashrc: Delete clutter
Delete all the unused portions of the bashrc, mostly aliases and
functions that were included as part of the example file I used,
and I forgot they existed. The reason behind deletion is to strive
for a cleaner dotfile setup - anything that ends up being missed
will be added back in.