Begin using Neovim
Change from vim to Neovim. VISUAL env variable and gitconfig editor are
both set to nvim.

Much of my vimrc (which remains in the vim/ dir for now) was written by
someone else - I didn't know it well because I didn't create it. The
neovim init.nvim is made with some inspiration but nothing is blidly
added.

There is still much to set up - autocompletion will probably be neovim
LSP, more plugins to setup, but this is a good start.