Initial nvim LSP setup
nvim-lspconfig is configured with lua, which is initiated by the
top-level init.vim. The namespace inside of the lua/ directory is
required with `lua require('namespace')` and this essentially calls the
init.lua within that directory. init.lua has the require statements for
each lua file in that namespace directory.

The initial nvim-lspconfig.lua file is copied from the example file in
the nvim-lspconfig repository. They will definitely be changed because
they're overwriting my regular leader commands.